+++ three/loaders/MD2Loader.d.ts
@@ -1,0 +1,18 @@
+/// <reference path="../three.d.ts" />
+
+declare module THREE {
+
+    export class MD2Loader {
+
+        constructor(manager: LoadingManager);
+
+        manager: LoadingManager;
+
+        ["constructor"]: typeof MD2Loader;
+
+        load: __internal.Loader.Loader2<Geometry>;
+        parse(): Geometry;
+
+    }
+
+}